+++ src/array/cpu/array_from_vector.cc
@@ -1,0 +1,13 @@
+
+namespace dgl {
+using runtime::NDArray;
+namespace aten {
+
+// TODO
+template<kDLDeviceType XPU, typename T>
+NDArray VecToNDArray(const std::vector<T> &vec, DLDataType dtype, DLContext ctx) {
+}
+
+};  // namespace aten
+
+};  // namespace dgl
